The initial commit of the nauci base entry script
The script can be used to create a list of users, add those users to a list of existing groups as well as a developer and usb group.

I have included a readme that is simply a copy and paste of the usage output from the script. I will improve on this in the future and will also provide a link to the docker hub repository.