Added shared volume support with POSIX ACL
The script now set's up the ACL rules of the attached volume so that
the user does not have to. For this to work, the file system must
support extended attributes so that they are attached directly to the
nodes otherwise the container will not have access to the ACL rules.

I have also added the dockerfile for the nauci base image.