Added traliing command support and ignore -s if not interactive
Any command that you place following the entrypoint optional
parameters will be evaluated at the end of the script.

If you pass in the -s and enter an interactive session, then any
trailing commands will be processed as soon as you exit the
interactive session.

If the script is not running interactively with a pseudoterminal
available then it will print a warning to stderr and ignore the -s
flag.